+++ integration/src/wallets/suite.ts
@@ -4,7 +4,7 @@ import {
 } from '@shapeshiftoss/hdwallet-core'
 
 export type InfoCreater = () => HDWalletInfo
-export type Creater = () => Promise<HDWallet>
+export type Creater = (type: any = null) => Promise<HDWallet>
 export type Getter = () => HDWallet
 export type Suite = (get: Getter) => void